@@ -1,5 +1,6 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+
+import { createRoot } from 'react-dom/client';
 import ReactGA from "react-ga4";
 import './index.css';
 import App from './App';
@@ -8,12 +9,15 @@ import * as serviceWorker from './serviceWorker';
 ReactGA.initialize('G-EQMR17KKJG');
 ReactGA.send("pageview");
 
-ReactDOM.render(
+const domNode =document.getElementById('root');
+const root = createRoot(
+ domNode 
+);
+root.render(
   <React.StrictMode>
     <App />
   </React.StrictMode>,
-  document.getElementById('root')
-);
+)
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
